fix bug "ERROR at NMAXTHREADS < nthreads"
For some reason the ar struct of the RS exceeds the number of allowed threads for some models (Isensee and Sobotta).
I did not understand hoiw this is usually handled. The numebr of threads is never explicitly limited to nMaxThreads in the usual Setups.
But this can be done manually with the arSetParallelThreads function.