consider condition-dependent input functions
This was implemented for dealing the original version of the Sobotta model.
Now, I rearanged the Sobotta model, such that the different inputs can be expressed as different parameter sets of the same input function.

This code is therefore not needed anymore for the benchmarks. But might be useful in the future.