fix bug in RSTemplate
For some models the dose repsonse conditions of the template and the RS were not merged in the same way.
The reason was that the merging mechanism is based on the observables (which are randomized in RS).
Now: for RS the merging is based only on the data.def file name